DCGAN/requirements.txt
matplotlib==3.10.0
numpy==2.2.2
python-dotenv==1.0.1
torch==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.0
